bug: store_tokens takes i64 but token counts are u64 — unchecked casts at every call site (#2182)
* bug: store_tokens takes i64 but token counts are u64 — unchecked casts at every call site

Task #2177
Agent: claude
Attempt: 1

* style: run cargo fmt on response_handler.rs


* fix: use checked u64->i64 conversion at SQL bind sites

Replace unchecked `as i64` casts with `i64::try_from(...).unwrap_or(i64::MAX)`
at both DB write boundaries in `store_tokens` and `complete_run`, preventing
silent sign corruption if token counts ever exceed i64::MAX.


---------

+++ src/engine/review.rs
@@ -1234,11 +1234,11 @@ pub(crate) async fn review_and_merge(
         input_tokens: agent_result_for_tokens
             .as_ref()
             .and_then(|r| r.input_tokens)
-            .unwrap_or(0) as i64,
+            .unwrap_or(0),
         output_tokens: agent_result_for_tokens
             .as_ref()
             .and_then(|r| r.output_tokens)
-            .unwrap_or(0) as i64,
+            .unwrap_or(0),
         total_cost_usd: agent_result_for_tokens
             .as_ref()
             .and_then(|r| r.cost_usd)
